# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Small Squaretable API
  description: |
    Enterprise-grade multi-tenant SaaS platform for AI chat and character management.

    ## Authentication

    Most endpoints require authentication via JWT Bearer tokens. Include the access token in the Authorization header:

    ```
    Authorization: Bearer <your_access_token>
    ```

    ## Error Handling

    All errors follow a standard format:

    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable error message",
        "details": {}
      },
      "meta": {
        "timestamp": "2026-02-04T12:00:00.000Z"
      }
    }
    ```

    ## Pagination

    List endpoints support pagination via query parameters:

    - `page`: Page number (default: 1)
    - `limit`: Items per page (default: 20, max: 100)
    - `sortBy`: Field to sort by
    - `sortOrder`: Sort order (asc/desc, default: desc)

    ## Caching

    Some endpoints use Redis caching. Check the `X-Cache` response header:
    - `HIT`: Response served from cache
    - `MISS`: Response served from database

  version: 0.1.0
  contact:
    name: Small Squaretable Team
    url: https://github.com/small-squaretable
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.small-squaretable.com
    description: Production server
...
paths:
...
  /api/v1/characters:
    post:
      tags:
        - Characters
      summary: Create character
      description: Creates a new character for the authenticated user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCharacterRequest'
      responses:
        '201':
          description: Character created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Character'
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      tags:
        - Characters
      summary: List user's characters
      description: Returns paginated list of the user's private characters
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Character list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedCharacters'
  /api/v1/characters/marketplace:
    get:
      tags:
        - Characters
      summary: Browse marketplace
      description: Returns paginated list of public characters (authentication optional)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Marketplace characters
          headers:
            X-Cache:
              description: Cache status (HIT or MISS)
              schema:
                type: string
                enum: [HIT, MISS]
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedCharacters'
  /api/v1/characters/search:
    get:
      tags:
        - Characters
      summary: Search characters
      description: Full-text search across public characters
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          description: Search query
          required: true
          schema:
            type: string
            minLength: 1
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
            example: "helper,coding"
        - name: sort
          in: query
          description: Sort order
          schema:
            type: string
            enum: [popular, newest, rating, relevance]
            default: relevance
        - name: filter
          in: query
          description: Additional filters
          schema:
            type: string
            enum: [all, safe, nsfw]
            default: safe
        - name: isNsfw
          in: query
          description: Include NSFW content
          schema:
            type: boolean
            default: false
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Search results
          headers:
            X-Cache:
              description: Cache status
              schema:
                type: string
                enum: [HIT, MISS]
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SearchResults'
  /api/v1/characters/{characterId}:
    get:
      tags:
        - Characters
      summary: Get character by ID
      description: Returns character details
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CharacterId'
      responses:
        '200':
          description: Character details
          headers:
            X-Cache:
              description: Cache status
              schema:
                type: string
                enum: [HIT, MISS]
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Character'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Characters
      summary: Update character
      description: Updates an existing character
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CharacterId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCharacterRequest'
      responses:
        '200':
          description: Character updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Character'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Characters
      summary: Delete character
      description: Permanently deletes a character
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CharacterId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/characters/{characterId}/publish:
    post:
      tags:
        - Characters
      summary: Publish character to marketplace
      description: Makes a character publicly available (requires Pro/Team plan)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CharacterId'
      responses:
        '200':
          description: Character published
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Character'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/characters/{characterId}/unpublish:
    post:
      tags:
        - Characters
      summary: Unpublish character
      description: Removes a character from the marketplace
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CharacterId'
      responses:
        '200':
          description: Character unpublished
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Character'
  /api/v1/characters/{characterId}/fork:
    post:
      tags:
        - Characters
      summary: Fork character
      description: Creates a copy of a public character for personal use
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CharacterId'
      responses:
        '201':
          description: Character forked
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Character'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/characters/{characterId}/ratings:
    post:
      tags:
        - Characters
      summary: Submit rating
      description: Rates a character (Pro/Team only)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CharacterId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingRequest'
      responses:
        '201':
          $ref: '#/components/responses/Success'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      tags:
        - Characters
      summary: Get character ratings
      description: Returns aggregated ratings for a character
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CharacterId'
      responses:
        '200':
          description: Ratings data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RatingResponse'

    put:
      tags:
        - Characters
      summary: Update rating
      description: Updates an existing rating
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CharacterId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingRequest'
      responses:
        '200':
          $ref: '#/components/responses/Success'

    delete:
      tags:
        - Characters
      summary: Delete rating
      description: Removes a user's rating
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CharacterId'
      responses:
        '200':
          $ref: '#/components/responses/Success'

  # ============================================================
  # Chat Endpoints
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (7 day expiration)
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sortBy
      in: query
      description: Field to sort by
      schema:
        type: string
    SortOrderParam:
      name: sortOrder
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    CharacterId:
      name: characterId
      in: path
      required: true
      description: Character ID
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
        error:
          type: object
          nullable: true
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
...
    Character:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 2000
          nullable: true
        avatarUrl:
          type: string
          format: uri
          maxLength: 2000
          nullable: true
        cardData:
          type: object
          additionalProperties: true
          description: SillyTavern character card data
        tags:
          type: array
          items:
            type: string
          maxItems: 20
        category:
          type: string
          maxLength: 50
          nullable: true
        isPublic:
          type: boolean
        isNsfw:
          type: boolean
        downloadCount:
          type: integer
          default: 0
        avgRating:
          type: number
          nullable: true
        ratingCount:
          type: integer
          default: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCharacterRequest:
      type: object
      required: [name, cardData]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        avatarUrl:
          type: string
          format: uri
          maxLength: 2000
        cardData:
          type: object
          description: SillyTavern character card data
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
        category:
          type: string
          maxLength: 50
        isNsfw:
          type: boolean
          default: false
    UpdateCharacterRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        avatarUrl:
          type: string
          format: uri
          maxLength: 2000
        cardData:
          type: object
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
        category:
          type: string
          maxLength: 50
        isNsfw:
          type: boolean
...
    PaginatedCharacters:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Character'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    SearchResults:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Character'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
        query:
          type: string
          description: Original search query

    # ============================================================
    # Rating Schemas
    # ============================================================
    RatingRequest:
      type: object
      required: [quality, creativity, interactivity, accuracy, entertainment]
      properties:
        quality:
          type: integer
          minimum: 1
          maximum: 5
          description: Quality rating (1-5 stars)
        creativity:
          type: integer
          minimum: 1
          maximum: 5
          description: Creativity rating (1-5 stars)
        interactivity:
          type: integer
          minimum: 1
          maximum: 5
          description: Interactivity rating (1-5 stars)
        accuracy:
          type: integer
          minimum: 1
          maximum: 5
          description: Accuracy rating (1-5 stars)
        entertainment:
          type: integer
          minimum: 1
          maximum: 5
          description: Entertainment value (1-5 stars)
        comment:
          type: string
          maxLength: 500
          description: Optional comment
    RatingResponse:
      type: object
      properties:
        characterId:
          type: string
          format: uuid
        avgQuality:
          type: number
        avgCreativity:
          type: number
        avgInteractivity:
          type: number
        avgAccuracy:
          type: number
        avgEntertainment:
          type: number
        overall:
          type: number
          description: Weighted average rating
        ratingCount:
          type: integer
        userRating:
          type: object
          nullable: true
          properties:
            quality:
              type: integer
            creativity:
              type: integer
            interactivity:
              type: integer
            accuracy:
              type: integer
            entertainment:
              type: integer
            comment:
              type: string

    # ============================================================
    # Chat Schemas
    # ============================================================
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: VALIDATION_ERROR
                  message:
                    type: string
                  details:
                    type: object
              meta:
                type: object
                properties:
                  timestamp:
                    type: string
                    format: date-time
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: FORBIDDEN
                  message:
                    type: string
                    example: Insufficient permissions for this operation
              meta:
                type: object
                properties:
                  timestamp:
                    type: string
                    format: date-time
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: NOT_FOUND
                  message:
                    type: string
                    example: The requested resource was not found
              meta:
                type: object
                properties:
                  timestamp:
                    type: string
                    format: date-time
...